# Conflicts:
#	lib/abstract_game/GameIDs.hpp
#	lib/scene/CMakeLists.txt
#	lib/scene/DashboardScene.cpp

+++ lib/abstract_game/GameRunThrough.hpp
@@ -16,14 +16,13 @@ namespace abstract_game {
  * This struct stores the information of a single runthrough of a game during a gamesession.
  */
     struct GameRunThrough {
-
         int gameRunThroughUID;
         size_t gameSessionUID;
-        long result;
+        double result;
         std::string resultUnit;
 
         GameRunThrough(size_t const &gameSessionUID, int const &gameRunThroughUID, std::string const &resultUnit,
-                       long const &result);
+                       double const &result);
 
     };
